+++ front_end_dev/ems/src/VolunteerHistory.css
@@ -1,0 +1,22 @@
+/* Volunteer History Table Styling */
+.volunteer-history-container {
+    text-align: center;
+    padding: 20px;
+}
+
+table {
+    width: 100%;
+    border-collapse: collapse;
+    margin-top: 20px;
+}
+
+th, td {
+    background-color: #8c9657;
+    padding: 10px;
+    border: 5px solid #444034;
+    text-align: left;
+}
+
+th {
+    font-weight: bold;
+}
